Split tests for stream and datagram socket options
