+++ comp/dogstatsd/listeners/uds_datagram_test.go
@@ -10,7 +10,11 @@
 package listeners
 
 import (
+	"net"
 	"testing"
+	"time"
+
+	"github.com/stretchr/testify/assert"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/dogstatsd/packets"
@@ -29,5 +33,49 @@ func TestStartStopUDSDatagramListener(t *testing.T) {
 }
 
 func TestUDSDatagramReceive(t *testing.T) {
-	testUDSReceive(t, udsDatagramListenerFactory, "unixgram")
+	socketPath := testSocketPath(t)
+
+	mockConfig := map[string]interface{}{}
+	mockConfig[socketPathConfKey("unixgram")] = socketPath
+	mockConfig["dogstatsd_origin_detection"] = false
+
+	var contents0 = []byte("daemon:666|g|#sometag1:somevalue1,sometag2:somevalue2")
+	var contents1 = []byte("daemon:999|g|#sometag1:somevalue1")
+
+	packetsChannel := make(chan packets.Packets)
+
+	config := fulfillDepsWithConfig(t, mockConfig)
+	s, err := udsDatagramListenerFactory(packetsChannel, newPacketPoolManagerUDS(config), config)
+	assert.Nil(t, err)
+	assert.NotNil(t, s)
+
+	s.Listen()
+	defer s.Stop()
+	conn, err := net.Dial("unixgram", socketPath)
+	assert.Nil(t, err)
+	defer conn.Close()
+
+	conn.Write(contents0)
+	conn.Write(contents1)
+
+	select {
+	case pkts := <-packetsChannel:
+		assert.Equal(t, 2, len(pkts))
+
+		packet := pkts[0]
+		assert.NotNil(t, packet)
+		assert.Equal(t, packet.Contents, contents0)
+		assert.Equal(t, packet.Origin, "")
+		assert.Equal(t, packet.Source, packets.UDS)
+
+		packet = pkts[1]
+		assert.NotNil(t, packet)
+		assert.Equal(t, packet.Contents, contents1)
+		assert.Equal(t, packet.Origin, "")
+		assert.Equal(t, packet.Source, packets.UDS)
+
+	case <-time.After(2 * time.Second):
+		assert.FailNow(t, "Timeout on receive channel")
+	}
+
 }
